start to map fields based on xml gaap templates

+++ StockApp/Models/CompanyModel.cs
@@ -9,7 +9,7 @@ namespace StockApp.Models
     public class CompanyModel
     {
         
-        private uint cik;
+        private UInt64 cik;
 
         
         private string ticker;
@@ -21,11 +21,12 @@ namespace StockApp.Models
         /// A TickersModel is a record representing a unique company in the SEC edgar database
         /// </summary>
         /// <param name="cik"> This is the CIK ID unique ID given to corps in the edgar database
-        /// The cik is a 10 digit identifier, we store this in the program as an unsigned integer for storage savings
-        /// This could fit into a tinyint type in sqlserver</param>
+        /// The cik is a 10 digit identifier, we store this in the program as an unsigned 64 bit integer to represent 10 digits
+        /// We could do better by maybe using a bitmap, but I will do that if I run into issues :)
+        /// </param>
         /// <param name="ticker"> This is the ticker for a company on a stock exchange</param>
         /// <param name="company_name"> This is the full name of a company</param>
-        public CompanyModel(uint cik, string ticker, string company_name)
+        public CompanyModel(UInt64 cik, string ticker, string company_name)
         {
             if (string.IsNullOrEmpty(ticker)) throw new ArgumentNullException(nameof(ticker));
             if (string.IsNullOrEmpty(company_name)) throw new ArgumentNullException(nameof(company_name));
@@ -38,7 +39,7 @@ namespace StockApp.Models
         /// <summary>
         /// CIK index is the unique identifier for a company in the edgar database
         /// </summary>
-        public uint CIK { get => cik; set => cik = value; }
+        public UInt64 CIK { get => cik; set => cik = value; }
 
         /// <summary>
         /// Ticker is the symbol of the company on its exchange
